backend/requirements.txt
amqp==5.1.1
asgiref==3.4.0
billiard==3.6.4.0
boto3==1.23.10
botocore==1.26.10
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.19
django-celery-beat==2.2.0
django-celery-results==2.0.1
idna==3.4
importlib-metadata==4.8.3
jmespath==0.10.0
jsonschema==2.6.0
kombu==5.2.4
Mako==1.1.6
MarkupSafe==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.4
requests==2.27.1
django-simple-captcha==0.5.17
django-timezone-field==4.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extensions==0.7.1
gitdb==4.0.9
GitPython==3.1.3
channels==2.3.1
django-simple-history==3.2.0
pycryptodome==3.17
openpyxl==3.1.1
django-cidrfield==0.2.0
django-import-export==3.1.0
netaddr==0.8.0
django-filter==22.1
numpy==1.23.5
Cython==0.29.33
wheel==0.37.1
pandas==1.1.5
pymongo==4.3.3
django-restql==0.15.3
user-agents==2.2.0
django-redis==5.2.0
paramiko==3.0.0
xmltodict==0.13.0
ncclient==0.6.13
ttp==0.9.2
nornir==3.3.0
nornir-jinja2==0.2.0
nornir-napalm==0.2.0
nornir-netmiko==0.1.2
nornir-utils==0.2.0
netmiko==3.4.0
channels-redis==4.0.0
PyDriller==2.1
supervisor==4.2.4
ipaddr==2.2.0
netaddr==0.8.0
flower==1.2.0
celery-once==3.0.1
pika~=1.3.1
aio-pika==9.1.2
prometheus-client==0.17.0
motor~=3.1.2
pydantic==1.10.8
gunicorn==20.1.0
uwsgi==2.0.21
daphne==2.3.0
schedule==1.2.0
drf-spectacular==0.26.2
py2neo==2021.2.3
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-fsm-log==3.1.0
textfsm==1.1.0